# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/login/byEmailPwd": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "邮箱加密码登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.LoginEmailPwdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/res.LoginInfoRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "req.LoginEmailPwdReq": {
            "type": "object",
            "required": [
                "email",
                "limit",
                "page",
                "pwd"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "limit": {
                    "description": "每页返回多少",
                    "type": "integer"
                },
                "page": {
                    "description": "页数",
                    "type": "integer"
                },
                "pwd": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "res.LoginInfoRes": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "token": {
                    "description": "登录token",
                    "type": "string"
                },
                "uid": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
        "res.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误code码",
                    "type": "integer"
                },
                "data": {
                    "description": "成功时返回的对象"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}